% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asthmaHarmonizer-package.R
\docType{package}
\name{asthmaHarmonizer-package}
\alias{asthmaHarmonizer}
\alias{asthmaHarmonizer-package}
\title{asthmaHarmonizer: Asthma Variable Harmonization Lookup}
\description{
A simple tool for looking up harmonized variable names and concepts for asthma-related research projects. It helps map raw variables to standardized concepts easily.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nzkermani/asthmaHarmonizer}
  \item Report bugs at \url{https://github.com/nzkermani/asthmaHarmonizer/issues}
}

}
\author{
\strong{Maintainer}: Nazanin Zounemat-Kermani \email{n.kermani@imperial.ac.uk}

}
\keyword{internal}
