% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_variable.R
\name{lookup_variable}
\alias{lookup_variable}
\title{Lookup Asthma Harmonization Info}
\usage{
lookup_variable(variable_name)
}
\arguments{
\item{variable_name}{A string. The variable you want to lookup.}
}
\value{
A tibble with harmonized information.
}
\description{
Given a variable name, returns its harmonization mapping based on the
internal harmonized_database.
}
\examples{
lookup_variable("age")
lookup_variable("sex")
}
