% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_added_variables.R
\name{view_added_variables}
\alias{view_added_variables}
\title{View Added Variables Log}
\usage{
view_added_variables()
}
\value{
A tibble with the log entries.
}
\description{
Displays all variables that were added via ChatGPT harmonization.
}
\examples{
view_added_variables()
}
