% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_variable.R
\name{lookup_variable}
\alias{lookup_variable}
\title{Lookup Asthma Harmonization Info (dynamic, LLM-enhanced)}
\usage{
lookup_variable(variable_name)
}
\arguments{
\item{variable_name}{A string. The variable you want to lookup.}
}
\value{
A tibble with harmonized information or triggers ChatGPT if not found.
}
\description{
Lookup Asthma Harmonization Info (dynamic, LLM-enhanced)
}
\examples{
if (interactive()) {
lookup_variable("age")
}
}
